combined formatted profile with stateful components
